move get_form_types & get_forms_for_user to utils
