+++ wagtail/contrib/forms/wagtail_hooks.py
@@ -3,9 +3,9 @@ from django.urls import reverse
 from django.utils.translation import ugettext_lazy as _
 
 from wagtail.admin.menu import MenuItem
-from wagtail.core import hooks
 from wagtail.contrib.forms import urls
-from wagtail.contrib.forms.models import get_forms_for_user
+from wagtail.contrib.forms.utils import get_forms_for_user
+from wagtail.core import hooks
 
 
 @hooks.register('register_admin_urls')
